Persist template cache across tests
This was introduced in d24e37db6e2866dfc45b5d76a7a604c377857161 [2018-06-17 :/Fix caches interfering with tests] solely because of CSRF tokens and the way `get_token` was monkeypatched, but we don’t use CSRF tokens anymore, so it was undocumented unnecessary code.

Incidentally saves test run time on the order of a couple of seconds.